chore: 继承base类
